+++ CHANGES.md
@@ -4,17 +4,23 @@
 
 ## To-Do
 
+- 증권사 Open API 샘플 프로그램 제작
 - Python 라이브러리 배포 환경 구성
 - Python 테스트 자동화 환경 구성
 - Node.js 라이브러리 배포 환경 구성
-- 개발 환경 구성 확인을 위한 Sample Program 제작
-- ESLint 환경 구성
-- Pretty 환경 구성
 
   
 
 ## Version History
 
+- 0.00.003 : ESLint와 Prettier 환경 구성
+  - Library
+    - eslint, prettier
+    - typescript-eslint
+    - eslint-config-prettier eslint-plugin-prettier
+  - VSCode extension
+    - dbaeumer.vscode-eslint
+    - esbenp.prettier-vscode
 - 0.00.002 : TypeScript 테스트 자동화 환경 구성
   - Library : [vitest](https://vitest.dev/guide/)
 - 0.00.001 : TypeScript 개발 환경 구성